backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
numpy>=2.2.0
pydantic==2.10.3
python-multipart==0.0.18
supabase==2.11.0
python-dotenv==1.0.1
opencv-python-headless>=4.11.0